Implemented a function to refine the query
We are aiming to use RAG for our chat, and hence, each time the user sends a query, we need to refine it depending on the chat history in order to get a query that can be used for similarity search. E.g. If the user asks the following question: 'How can I fix it?' and before that he was talking about binge eating, this function will turn his query into 'How can I fix binge eating?' so that we can retrieve relevant documents from our knowledge base to give tips for binge eating