Updated query refinement function
Given that Gemini has a limited context window, we can't pass the entire chat history in the prompt if the chat history is huge. What I am doing here is to only take the last 15 message exchanges between the AI and the user in order to refine the query. Which should be enough anyway!